pyproject.toml
[tool.poetry]
name = "interactive-ink"
version = "0.1.0"
description = ""
authors = ["nwilson314 <nsw.wilson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
black = "^24.8.0"
fastapi = {extras = ["standard"], version = "^0.115.0"}
llama-index = "^0.11.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
